app/backend/requirements.txt
Django==4.2.7
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
torch>=2.0.0
mne @ git+https://github.com/devparikh0506/mne-python.git@fc98279b3b3ee9c76b3820d40e16d4c1b9dba570
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

